fix: move default vb logic to app data
